Backend/requirements.txt
    fastapi[all]>=0.100.0  
    sqlalchemy>=2.0.0
    psycopg2-binary>=2.9.0 
    pydantic-settings>=2.0.0
    python-dotenv>=1.0.0
    passlib[bcrypt]>=1.7.4 
    python-jose[cryptography]>=3.3.0 

    azure-cognitiveservices-vision-computervision>=0.9.0
    msrest>=0.7.1

    openai>=1.0.0
    google-generativeai>=0.3.0
    langchain>=0.0.157
    langchain-chroma>=0.1.0
    sentence-transformers>=2.2.2
    langchain-google-genai>=0.0.5
    langchain-community>=0.0.10
    chromadb>=0.4.0
    python-multipart>=0.0.6  
    PyPDF2>=3.0.0 

    python-slugify>=8.0.0  
    python-magic>=0.4.27  
    boto3>=1.24.0
    langchain-openai>=0.0.157

